Split scene.rs into modular scene files
Remove the large monolithic crates/helio/src/scene.rs and split its contents into a scene module tree for better organization and maintainability. New files include camera.rs, core.rs, errors.rs, helpers.rs, types.rs, mod.rs, grouped submodules (groups/{membership,transforms,visibility}.rs), object operations (objects/{insert,update,remove,rebuild,mod}.rs), resource management (resources/{lights,materials,meshes,textures,mod}.rs) and virtual_geometry support (virtual_geometry/{meshes,objects,rebuild,mod}.rs). This refactor reorganizes the same scene functionality into focused files without changing external behavior (logical split only).
